Until now this is my finalized code of PositiveBigInt Datatype
--> It took me 2 whole days of thinking, coding and
debugging to write this whole code, 8hrs each day.

--> I used my "Naive Approach" and wrote this code
from scratch without any help of any AI or Google.

--> Although using AI and Google is not a problem
at all as a coder, I know, and it is not something
to consider a big of an achievement as a coder
if you code something without any help of any AI or
Google, because honestly as coder I know and I have
noticed that the efficiency gets really improved
when I take help of AIs and Google while coding
and most of the times I do use both of these things.

--> But this time I didn't use any of these because
I just wanted to check that how much of OOPS
concepts do I still remember and I also wanted to
check that can I can really implement at least
the primary school level of mathematics logic of
addition and multiplication on my own, and I have
to say this, that I haven't really revised OOPS in
a while, but while writing this code I saw that I
still have a nice grip on OOPS, and I was able to
code basic mathematics logics. Just because of these
things, I am proud of myself a little bit today.